+++ example/lib/main.dart
@@ -100,7 +100,10 @@ class _DemoState extends State<Demo> {
                     padding: const EdgeInsets.symmetric(horizontal: 8.0),
                     child: TextField(
                       prefix: Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 4.0),
+                        padding: const EdgeInsets.symmetric(
+                          horizontal: 4.0,
+                          vertical: 2.0,
+                        ),
                         child: Icon(CupertinoIcons.search),
                       ),
                       placeholder: 'Type some text here',
@@ -113,20 +116,30 @@ class _DemoState extends State<Demo> {
                     ),
                   ),
                   const SizedBox(height: 20),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 8.0),
-                    child: TextField.borderless(
-                      prefix: Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 4.0),
-                        child: Icon(CupertinoIcons.search),
-                      ),
-                      placeholder: 'Type some text here',
+                  Label(
+                    icon: Icon(
+                      CupertinoIcons.tag,
+                      color: CupertinoColors.activeBlue,
+                    ),
+                    text: SelectableText('A borderless textfield: '),
+                    child: Expanded(
+                      child: Padding(
+                        padding: const EdgeInsets.symmetric(horizontal: 8.0),
+                        child: TextField.borderless(
+                          prefix: Padding(
+                            padding:
+                                const EdgeInsets.symmetric(horizontal: 4.0),
+                            child: Icon(CupertinoIcons.search),
+                          ),
+                          placeholder: 'Type some text here',
 
-                      /// If both suffix and clear button mode is provided,
-                      /// suffix will override the clear button.
-                      suffix: Text('SUFFIX'),
-                      // clearButtonMode: OverlayVisibilityMode.always,
-                      maxLines: null,
+                          /// If both suffix and clear button mode is provided,
+                          /// suffix will override the clear button.
+                          suffix: Text('SUFFIX'),
+                          // clearButtonMode: OverlayVisibilityMode.always,
+                          maxLines: null,
+                        ),
+                      ),
                     ),
                   ),
                 ],