+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.1]
+* Implemented `Label` ([#61](https://github.com/GroovinChip/macos_ui/issues/61))
+* Capacity Indicator now works as expected ([#49](https://github.com/GroovinChip/macos_ui/issues/49))
+* Clear button is now aligned to text ([#82](https://github.com/GroovinChip/macos_ui/issues/82))
+
 ## [0.1.0]
 * pub.dev release
 
@@ -12,7 +17,7 @@
 * Implement `TextField`
 
 ## [0.0.10]
-* Revamp `Scaffold` [#26] (https://github.com/GroovinChip/macos_ui/issues/26)
+* Revamp `Scaffold` [#26](https://github.com/GroovinChip/macos_ui/issues/26)
 
 ## [0.0.9+1]
 * `CapacityIndicator` colors can now be set on its constructor